Refactor imports to use shared utilities and add TypeScript configuration
