+++ packages/shared/package.json
@@ -3,10 +3,21 @@
     "version": "0.0.0",
     "private": true,
     "exports": {
-        ".": "./src/index.tsx",
-        "./*": "./src/*.tsx"
+        ".": {
+            "types": "./src/index.ts",
+            "default": "./src/index.ts"
+        },
+        "./utils": {
+            "types": "./src/utils/index.ts",
+            "default": "./src/utils/index.ts"
+        },
+        "./*": {
+            "types": "./src/*.tsx",
+            "default": "./src/*.tsx"
+        }
     },
-    "main": "./src/index.tsx",
+    "main": "./src/index.ts",
+    "types": "./src/index.ts",
     "scripts": {
         "lint": "eslint . --max-warnings 0",
         "generate:component": "turbo gen react-component",
@@ -31,3 +42,4 @@
         "tailwindcss": "^4.1.7"
     }
 }
+
